抽取提示词当前时间测试断言
