+++ tests/prompt-assertions.ts
@@ -1,0 +1,10 @@
+import assert from 'node:assert/strict';
+
+export function assertPromptCurrentTimeHasGanzhiCalendar(prompt: string) {
+  const currentTimeSection = prompt.match(/^【当前时间】\n([\s\S]*?)(?=\n【)/m)?.[1] ?? '';
+
+  assert.match(currentTimeSection, /^公历：\d{4}年\d{1,2}月\d{1,2}日 \d{1,2}时\d{1,2}分/m);
+  assert.match(currentTimeSection, /^农历：.+[子丑寅卯辰巳午未申酉戌亥]时$/m);
+  assert.match(currentTimeSection, /^干支历：.+年 .+月 .+日 .+时$/m);
+  assert.match(currentTimeSection, /^当前节气：.+/m);
+}
